code/backend/requirements.txt
Flask
Werkzeug
flask_bootstrap4
flask_compress
flask_reuploaded
flask_cors
flask_limiter
keras
numpy
matplotlib
